81 - finish moving functions into individual files
