+++ src/mnemonicToSeedHex.js
@@ -1,0 +1,7 @@
+var mnemonicToSeed = require('./mnemonicToSeed')
+
+function mnemonicToSeedHex (mnemonic, password) {
+  return mnemonicToSeed(mnemonic, password).toString('hex')
+}
+
+module.exports = mnemonicToSeedHex
