+++ src/salt.js
@@ -1,0 +1,5 @@
+function salt (password) {
+  return 'mnemonic' + (password || '')
+}
+
+module.exports = salt
